refactor: organize txoutput components into their own files
