+++ packages/app-explorer/src/systems/Transaction/component/TxOutput/types.ts
@@ -1,0 +1,6 @@
+import type { GQLTransactionOutputFragment } from '@fuel-explorer/graphql';
+import type { CardProps } from '@fuels/ui';
+
+export type TxOutputProps<T = GQLTransactionOutputFragment> = CardProps & {
+  output: T;
+};
